+++ tox.ini
@@ -13,6 +13,7 @@ deps =
     nose-timer
     coverage<4
     coveralls
+    docopt
 commands =
     nosetests --with-coverage --with-timer --cover-package=tqdm --ignore-files="tests_perf\.py" -d -v tqdm/
     coveralls
@@ -37,5 +38,6 @@ commands =
 deps =
     nose
     nose-timer
+    docopt
 commands =
     nosetests --with-timer tqdm/tests/tests_perf.py -d -v
